define shirotuljabar_mat
unsigned int num_rows - represents the number of rows of the matrix. 0 is not an acceptable value;
unsigned int num_cols - represents the number of columns of the matrix. 0 is not an acceptable value;
double **data - is “multi-dimensional array” that stores data in rows and columns;
int is_square - is a "boolean" value that determines if the matrix is square (has the same number of rows and columns) or not.